refactor: move file system code to data/fs
