+++ core/CMakeLists.txt
@@ -40,10 +40,10 @@ set(CUBOS_CORE_SOURCE
     "src/cubos/core/data/binary_serializer.cpp"
     "src/cubos/core/data/binary_deserializer.cpp"
     "src/cubos/core/data/package.cpp"
-    "src/cubos/core/data/file.cpp"
-    "src/cubos/core/data/file_system.cpp"
-    "src/cubos/core/data/standard_archive.cpp"
-    "src/cubos/core/data/embedded_archive.cpp"
+    "src/cubos/core/data/fs/file.cpp"
+    "src/cubos/core/data/fs/file_system.cpp"
+    "src/cubos/core/data/fs/standard_archive.cpp"
+    "src/cubos/core/data/fs/embedded_archive.cpp"
     "src/cubos/core/data/qb_parser.cpp"
     "src/cubos/core/data/context.cpp"
 
@@ -103,12 +103,12 @@ set(CUBOS_CORE_INCLUDE
     "include/cubos/core/data/binary_deserializer.hpp"
     "include/cubos/core/data/serialization_map.hpp"
     "include/cubos/core/data/package.hpp"
-    "include/cubos/core/data/file.hpp"
-    "include/cubos/core/data/file_stream.hpp"
-    "include/cubos/core/data/file_system.hpp"
-    "include/cubos/core/data/archive.hpp"
-    "include/cubos/core/data/standard_archive.hpp"
-    "include/cubos/core/data/embedded_archive.hpp"
+    "include/cubos/core/data/fs/file.hpp"
+    "include/cubos/core/data/fs/file_stream.hpp"
+    "include/cubos/core/data/fs/file_system.hpp"
+    "include/cubos/core/data/fs/archive.hpp"
+    "include/cubos/core/data/fs/standard_archive.hpp"
+    "include/cubos/core/data/fs/embedded_archive.hpp"
     "include/cubos/core/data/qb_parser.hpp"
     "include/cubos/core/data/context.hpp"
