+++ core/include/cubos/core/data/fs/embedded_archive.hpp
@@ -6,7 +6,7 @@
 #include <map>
 #include <vector>
 
-#include <cubos/core/data/archive.hpp>
+#include <cubos/core/data/fs/archive.hpp>
 
 namespace cubos::core::data
 {
