--- core/samples/data/embedded_archive/assets/bar/baz.txt
@@ -1,1 +1,0 @@
-baz
