--- core/samples/data/embedded_archive/assets/foo.txt
@@ -1,1 +1,0 @@
-foo
