+++ core/samples/data/fs/embedded_archive/assets/bar/baz.txt
@@ -1,0 +1,1 @@
+baz
