+++ core/samples/data/fs/embedded_archive/assets/foo.txt
@@ -1,0 +1,1 @@
+foo
