+++ core/src/cubos/core/data/fs/standard_archive.cpp
@@ -1,5 +1,5 @@
-#include <cubos/core/data/file_stream.hpp>
-#include <cubos/core/data/standard_archive.hpp>
+#include <cubos/core/data/fs/file_stream.hpp>
+#include <cubos/core/data/fs/standard_archive.hpp>
 #include <cubos/core/log.hpp>
 #include <cubos/core/memory/standard_stream.hpp>
