+++ engine/samples/assets/json/main.cpp
@@ -10,7 +10,7 @@
 
 #include <vector>
 
-#include <cubos/core/data/file_system.hpp>
+#include <cubos/core/data/fs/file_system.hpp>
 #include <cubos/core/settings.hpp>
 
 #include <cubos/engine/assets/bridges/json.hpp>
