+++ engine/src/cubos/engine/assets/assets.cpp
@@ -1,7 +1,7 @@
 #include <utility>
 
 #include <cubos/core/data/debug_serializer.hpp>
-#include <cubos/core/data/file_system.hpp>
+#include <cubos/core/data/fs/file_system.hpp>
 #include <cubos/core/data/json_deserializer.hpp>
 #include <cubos/core/data/json_serializer.hpp>
 #include <cubos/core/log.hpp>
