+++ engine/src/cubos/engine/file_settings/plugin.cpp
@@ -1,6 +1,6 @@
-#include <cubos/core/data/file_system.hpp>
+#include <cubos/core/data/fs/file_system.hpp>
 #include <cubos/core/data/json_deserializer.hpp>
-#include <cubos/core/data/standard_archive.hpp>
+#include <cubos/core/data/fs/standard_archive.hpp>
 #include <cubos/core/settings.hpp>
 
 #include <cubos/engine/file_settings/plugin.hpp>
